+++ packages/orq-rc/src/models/operations/updatebudget.ts
@@ -543,7 +543,7 @@ export const UpdateBudgetResponseBody$inboundSchema: z.ZodType<
   created: z.string().datetime({ offset: true }).transform(v => new Date(v))
     .optional(),
   updated: z.string().datetime({ offset: true }).default(
-    "2025-09-04T13:30:49.922Z",
+    "2025-09-04T15:00:50.847Z",
   ).transform(v => new Date(v)),
 }).transform((v) => {
   return remap$(v, {
@@ -581,7 +581,7 @@ export const UpdateBudgetResponseBody$outboundSchema: z.ZodType<
   isActive: z.boolean(),
   consumption: z.lazy(() => UpdateBudgetConsumption$outboundSchema).optional(),
   created: z.date().transform(v => v.toISOString()).optional(),
-  updated: z.date().default(() => new Date("2025-09-04T13:30:49.922Z"))
+  updated: z.date().default(() => new Date("2025-09-04T15:00:50.847Z"))
     .transform(v => v.toISOString()),
 }).transform((v) => {
   return remap$(v, {