+++ packages/orq-rc/src/models/operations/updatedataset.ts
@@ -302,7 +302,7 @@ export const UpdateDatasetResponseBody$inboundSchema: z.ZodType<
   created: z.string().datetime({ offset: true }).transform(v => new Date(v))
     .optional(),
   updated: z.string().datetime({ offset: true }).default(
-    "2025-09-04T13:30:49.922Z",
+    "2025-09-04T15:00:50.847Z",
   ).transform(v => new Date(v)),
 }).transform((v) => {
   return remap$(v, {
@@ -347,7 +347,7 @@ export const UpdateDatasetResponseBody$outboundSchema: z.ZodType<
   parentId: z.string().optional(),
   version: z.string().optional(),
   created: z.date().transform(v => v.toISOString()).optional(),
-  updated: z.date().default(() => new Date("2025-09-04T13:30:49.922Z"))
+  updated: z.date().default(() => new Date("2025-09-04T15:00:50.847Z"))
     .transform(v => v.toISOString()),
 }).transform((v) => {
   return remap$(v, {