+++ packages/orq-rc/RELEASES.md
@@ -3449,3 +3449,13 @@ Based on:
 - [typescript v3.12.0-rc.39] packages/orq-rc
 ### Releases
 - [NPM v3.12.0-rc.39] https://www.npmjs.com/package/@orq-ai/node/v/3.12.0-rc.39 - packages/orq-rc
+
+## 2025-09-04 15:02:39
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.609.0 (2.692.0) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v3.12.0-rc.40] packages/orq-rc
+### Releases
+- [NPM v3.12.0-rc.40] https://www.npmjs.com/package/@orq-ai/node/v/3.12.0-rc.40 - packages/orq-rc